requirements.txt
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0
openai>=1.0.0
plyer>=2.0.0